refactor(multi-select): move key handler inside the form control for easier refactoring to a multi-select control
