+++ src/components/multi-select/multi-select.jsx
@@ -1,6 +1,4 @@
 import React, {
-  createRef,
-  useEffect,
   useImperativeHandle,
   useRef,
   useState,
@@ -8,20 +6,31 @@ import React, {
 } from 'react';
 import PropTypes from 'prop-types';
 import FormControl from '../form-control/form-control.jsx';
-import FormControlIcons from '../form-control-icons/form-control-icons.jsx';
-import Icon from '../icon/icon.jsx';
-import IconButton from '../icon-button/icon-button.jsx';
-import iconCaretDown from '../../svgs/caret-down.svg';
-import iconClear from '../../svgs/clear.svg';
-import Listbox from '../listbox/listbox.jsx';
-import ListOption from '../list-option/list-option.jsx';
-import Loader from '../loader/loader.jsx';
-import NoOptions from '../no-options/no-options.jsx';
-import TagList from '../tag-list/tag-list.jsx';
-import Tag from '../tag/tag.jsx';
-import useDropdownClose from '../../hooks/use-dropdown-close.js';
+import MultiSelectControl from '../control/multi-select.jsx';
 import styles from './multi-select.css';
 import useForwardedRef from '../../hooks/use-forwarded-ref.js';
+
+function spread(ref1, ref2) {
+  const newObject = {};
+
+  Object.keys(ref1.current).reduce((acc, key) => (
+    Object.defineProperty(acc, key, {
+      get() { return ref1.current[key]; },
+      configurable: true,
+      enumerable: true,
+    })
+  ), newObject);
+
+  Object.keys(ref2.current).reduce((acc, key) => (
+    Object.defineProperty(acc, key, {
+      get() { return ref2.current[key]; },
+      configurable: true,
+      enumerable: true,
+    })
+  ), newObject);
+
+  return newObject;
+}
 
 function getFormControlChildrenContainerClassName(readOnly, validationMessage) {
   if (readOnly) {
@@ -36,189 +45,33 @@ function getFormControlChildrenContainerClassName(readOnly, validationMessage) {
 }
 
 const MultiSelect = forwardRef(function MultiSelect(props, ref) {
-  const [autocompleteValue, setAutocompleteValue] = useState('');
-  const [expanded, setExpanded] = useState(false);
   const selfRef = useForwardedRef(ref);
   const [validationMessage, setValidationMessage] = useState(props.validationMessage);
-  const [value, setValue] = useState(props.value || []);
-  const valueRefs = value.map(() => createRef());
-  const visibleOptions = getVisibleOptions();
-  const visibleOptionsRefs = visibleOptions.map(() => createRef());
 
   const ids = {
-    emptyMessage: `${props.id}-empty`,
     label: `${props.id}-label`,
-    listbox: `${props.id}-listbox`,
     textbox: `${props.id}-autocomplete`,
-    tooltip: `${props.id}-autocomplete-tooltip`,
-    validation: `${props.id}Hint`,
   };
   const refs = {
-    autocomplete: useRef(),
     control: useRef(),
-    wrapper: useRef(),
+    multiSelect: useRef(),
   };
-
   const classNames = {
     container: styles.container,
     formControlChildrenContainer: getFormControlChildrenContainerClassName(props.readOnly, validationMessage),
-    iconClear: styles['icon-clear'],
-    iconsContainer: styles['icon-container'],
     input: props.readOnly ? styles['input-readonly'] : styles.input,
-    noOptionsContainer: styles['no-options'],
-    popupContainer: styles['popup-container'],
     tagList: styles['tag-list'],
     ...props.classNames,
   };
 
-  const dropdownContents = () => {
-    if (!expanded) {
-      return undefined;
-    }
-
-    if (props.showLoader) {
-      return (
-        <Listbox
-          className={classNames.popupContainer}
-          id={ids.listbox}
-          labelledBy={ids.label}
-          refs={[]}
-        >
-          {() => <Loader inline />}
-        </Listbox>
-      );
-    }
-
-    if (visibleOptions.length === 0) {
-      return (<NoOptions className={classNames.noOptionsContainer} id={ids.emptyMessage} />);
-    }
-
-    return (
-      <Listbox
-        className={classNames.popupContainer}
-        id={ids.listbox}
-        labelledBy={ids.label}
-        refs={visibleOptionsRefs}
-      >
-        {() => (props.listboxChildren ?
-          props.listboxChildren(visibleOptions, visibleOptionsRefs, onOptionSelect) :
-          visibleOptions.map((option, index) => (
-            <ListOption
-              key={`${props.id}-${props.optionIDGetter(option)}`}
-              onSelect={onOptionSelect}
-              ref={visibleOptionsRefs[index]}
-              value={props.optionIDGetter(option)}
-            >
-              {props.optionLabelGetter(option)}
-            </ListOption>
-          ))
-        )}
-      </Listbox>
-    );
-  };
-
-  function getVisibleOptions() {
-    const filteredOptions = props.options.filter(option => !value.some((val) => {
-      return props.optionIDGetter(val) === props.optionIDGetter(option);
-    }));
-
-    if (!props.filterOptions) {
-      return filteredOptions;
-    }
-
-    return filteredOptions
-      .filter(option => props.optionLabelGetter(option).toLowerCase().includes(autocompleteValue.toLowerCase()));
+  function onInvalid(event) {
+    setValidationMessage(event.detail.validationMessage);
   }
 
-  function onAutocompleteBlur() {
-    if (!refs.autocomplete.current) return;
-
-    refs.autocomplete.current.setCustomValidity('');
-    if (!refs.autocomplete.current.validity.valid) {
-      setValidationMessage(refs.autocomplete.current.validationMessage);
-    } else {
-      setValidationMessage(props.validationMessage || '');
-    }
-  }
-
-  function onAutocompleteChange(event) {
-    setExpanded(true);
-    setAutocompleteValue(event.target.value);
-  }
-
-  function onDropdownClose() {
-    setExpanded(false);
-    setAutocompleteValue('');
-  }
-
-  function onOptionRemove(event) {
-    const newValue = value.filter((option, index) => (
-      valueRefs[index].current !== event.target
-    ));
-    setValue(newValue);
-  }
-
-  function onOptionsClear(event) {
-    event.preventDefault();
-
-    if (!props.readOnly) {
-      setValue([]);
-      setExpanded(false);
-      refs.autocomplete.current.focus();
-    }
-  }
-
-  function onOptionSelect(event) {
-    const selectedOptionIndex = visibleOptionsRefs.findIndex(optionRef => optionRef === event.target);
-    const selectedOption = visibleOptions[selectedOptionIndex];
-    setExpanded(false);
-    const newValue = [...value, selectedOption]
-      .sort((a, b) => props.optionIDGetter(a) - props.optionIDGetter(b));
-    setValue(newValue);
-    setAutocompleteValue('');
-    refs.autocomplete.current.focus();
-  }
-
-  function onClick(event) {
-    if (event.defaultPrevented) return;
-    if (props.readOnly) return;
-
-    setExpanded(true);
-  }
-
-  function onKeyDown(event) {
-    switch (event.key) {
-      case 'Escape':
-        setExpanded(false);
-        break;
-      default:
-    }
-  }
-
-  useDropdownClose(refs.control, expanded, onDropdownClose);
-
-  useEffect(() => {
-    setValidationMessage(props.validationMessage);
-  }, [props.validationMessage]);
-
-  useEffect(() => {
-    setValue(props.value);
-  }, [props.value.map(val => JSON.stringify(val)).join(',')]);
-
-  useEffect(() => {
-    props.onChange({ target: selfRef.current });
-  }, [value.length]);
-
-  useImperativeHandle(selfRef, () => ({
-    ...refs.control.current,
-    get dirty() {
-      return props.value.map(val => JSON.stringify(val)).join(',') !== this.value.map(val => JSON.stringify(val)).join(',');
-    },
-    id: props.id,
-    get value() {
-      return value;
-    },
-  }));
+  useImperativeHandle(selfRef, () => spread(
+    refs.control,
+    refs.multiSelect,
+  ));
 
   return (
     <FormControl
@@ -230,80 +83,34 @@ const MultiSelect = forwardRef(function MultiSelect(props, ref) {
       ref={refs.control}
       required={props.required}
       tooltip={props.tooltip}
-      validationMessage={validationMessage || ''}
+      validationMessage={validationMessage}
     >
-      <div
-        role="presentation"
-        className={classNames.formControlChildrenContainer}
-        onClick={onClick}
-        onKeyDown={onKeyDown}
-      >
-        <TagList
-          className={classNames.tagList}
-          labelledBy={ids.label}
-          refs={valueRefs}
-        >
-          {value.length !== 0 &&
-          props.tagChildren ?
-            props.tagChildren(value, valueRefs, onOptionRemove) :
-            value.map((val, index) => (
-              <Tag
-                defaultActive={index === 0}
-                id={`${props.id}-option-${props.optionIDGetter(val)}`}
-                key={props.optionIDGetter(val)}
-                onRemove={onOptionRemove}
-                readOnly={props.readOnly}
-                ref={valueRefs[index]}
-              >
-                {props.optionLabelGetter(val)}
-              </Tag>
-            ))
-          }
-          <input
-            aria-autocomplete="list"
-            aria-controls={ids.listbox}
-            aria-describedby={`${ids.emptyMessage} ${ids.tooltip}`}
-            aria-expanded={expanded}
-            aria-haspopup="listbox"
-            autoComplete="off"
-            role="combobox"
-            className={classNames.input}
-            id={ids.textbox}
-            onBlur={onAutocompleteBlur}
-            onChange={onAutocompleteChange}
-            onInput={props.onInput}
-            placeholder={value.length === 0 ? props.placeholder : undefined}
-            readOnly={props.readOnly}
-            required={props.required ? value.length === 0 : false}
-            ref={refs.autocomplete}
-            value={autocompleteValue}
-          />
-        </TagList>
-        <FormControlIcons
-          className={classNames.iconsContainer}
-          validationMessage={validationMessage}
-          validationMessageId={ids.validation}
-        >
-          {(!props.readOnly && value.length > 0) && (
-            <IconButton
-              icon={iconClear}
-              label={`Remove all selected options on ${props.label}`}
-              onPress={onOptionsClear}
-              className={classNames.iconClear}
-            />
-          )}
-          {props.readOnly ? (
-            <Icon icon={iconCaretDown} label={`Opening ${props.label} options disabled while read only`} />
-          ) : (
-            <IconButton
-              icon={iconCaretDown}
-              label={`Open ${props.label} options`}
-              onPress={onClick}
-            />
-          )}
-        </FormControlIcons>
-      </div>
-      {dropdownContents()}
+      <MultiSelectControl
+        classNames={{
+          formControlChildrenContainer: classNames.formControlChildrenContainer,
+          input: classNames.input,
+          tagList: classNames.tagList,
+        }}
+        containerRef={refs.control}
+        filterOptions={props.filterOptions}
+        id={props.id}
+        label={props.label}
+        listboxChildren={props.listboxChildren}
+        onChange={props.onChange}
+        onInput={props.onInput}
+        onInvalid={onInvalid}
+        options={props.options}
+        optionIDGetter={props.optionIDGetter}
+        optionLabelGetter={props.optionLabelGetter}
+        placeholder={props.placeholder}
+        readOnly={props.readOnly}
+        ref={refs.multiSelect}
+        required={props.required}
+        showLoader={props.showLoader}
+        tagChildren={props.tagChildren}
+        validationMessage={props.validationMessage}
+        value={props.value}
+      />
     </FormControl>
   );
 });
@@ -330,7 +137,7 @@ MultiSelect.propTypes = {
   placeholder: PropTypes.string,
   readOnly: PropTypes.bool,
   required: PropTypes.bool,
-  showLoader: PropTypes.bool, // eslint-disable-line react/no-unused-prop-types
+  showLoader: PropTypes.bool,
   tagChildren: PropTypes.func,
   tooltip: PropTypes.string,
   validationMessage: PropTypes.string,
